OPIK-153 refactor
